+++ uastreamer/src/components/mod.rs
@@ -1,10 +1,13 @@
 /// Holds everything related to the TCP Communication Channel
+#[deprecated]
 pub mod control;
 
 /// Holds all things streamer related
 //pub mod streamer;
 
+#[deprecated]
 pub mod cpal;
 
+#[deprecated]
 pub mod udp;
 pub mod tokio;
